Fix page leak when restoring a savepoint
Restoring a savepoint dropped processed transactions from the free tree.
However, this could lead to pages being leaked, if page A was allocated,
then added to the freed tree, then a savepoint was captured, and then
later A was freed a allocated again. When the savepoint was restored, A
would remain allocated but also be removed from the freed tree