Make all I/O errors fatal. Fixes potential corruption
If a transient I/O error occurred during the updates to a btree, but
then I/O errors did not occur when the Database was dropped. It was
possible for the file to be flushed with a clean recovery flag and
broken transaction, leading to corruption.

After an I/O all operations that touch the StorageBackend will now
return StorageError::PreviousIo