Major performance improvement to compact()
compact() previously moved pages down to lower indices in a semi-random
and optimistic way. However, this took O(N^2) time in the size of the
database.

This changes the compaction algorithm to identify the pages at the end
of the database file which can be moved, and only move those pages. This
is done iteratively until no more pages can be moved