Add set_quick_repair()
This adds a new quick-repair mode, which gives instant recovery after a
crash at the cost of slower commits.

To make this work, each commit with quick-repair enabled needs to save
the allocator state somewhere. We can't use the region headers, because
we'd be overwriting them in place; we might crash partway through the
overwrite, and then we'd need a full repair. So we instead save the
allocator state to a new table in the system tree. Writing to the table
is slightly tricky, because it needs to be done without allocating (see
below), but other than that it's a perfectly ordinary transactional write
with all the usual guarantees.

The other requirement to avoid full repair is knowing whether the last
transaction used 2-phase commit. For this, we add a new two_phase_commit
bit to the god byte, which is always updated atomically along with
swapping the primary bit. Old redb versions will ignore the new flag
when reading and clear it when writing, which is exactly what we want.

This turns out to also fix a longstanding bug where 2-phase commit hasn't
been providing any security benefit at all. The checksum forgery attack
described in the documentation for 1-phase commit actually works equally
well against 2-phase commit! The problem is that even though 2-phase
commit guarantees the primary is valid, redb ignores the primary flag
when repairing. It always picks whichever commit slot is newer, as long
as the checksum is valid. So if you crash partway through a commit,
it'll try to recover using the partially-written secondary rather than
the fully-written primary, regardless of the commit strategy.

The fix for this is exactly the two_phase_commit bit described above.
After a crash, we check whether the last transaction used 2-phase commit;
if so, we only look at the primary (which is guaranteed to be valid) and
ignore the secondary. Quick-repair needs this check anyway for safety,
so we get the bug fix for free.

To write to the allocator state table without allocating, I've introduced
a new insert_inplace() function. It's similar to insert_reserve(), but
more general and maybe simpler. To use it, you have to first do an
ordinary insert() with your desired key and a value of the appropriate
length; then later in the same transaction you can call insert_inplace()
to replace the value with a new one. Unlike insert_reserve(), this works
with values that don't implement MutInPlaceValue, and it lets you hold
multiple reservations simultaneously.

insert_inplace() could be safely exposed to users, but I don't think
there's any reason to. Since it doesn't give you a mutable reference,
there's no benefit over insert() unless you're storing data that cares
about its own position in the database. So for now it's private, and I
haven't bothered making a new error type for it; it just panics if you
don't satisfy the preconditions.

The fuzzer is perfect for testing quick-repair, because it can simulate
a crash, reopen the database (using quick-repair if possible), and then
verify that the resulting allocator state exactly matches what would
happen if it ran a full repair. I've modified the fuzzer to generate
quick-repair commits in addition to ordinary commits.