Fix several crashes in repair from corrupted pages
During repair, it's normal to encounter a partially-written tree, which
can contain arbitrary invalid data. We can't crash when that happens;
we need to gracefully fail so the repair can try the other commit slot
instead.

This first batch of fixes only covers the crashes most likely to occur in
practice: the ones that can happen on a 64-bit machine, without malicious
input.