Move commit() out of do_repair()
Instead of actually making a commit, do_repair() should just return the
new roots that would be committed. This makes check_integrity() much
faster, since it no longer needs to verify the checksums twice, and
it means commit() can always trim the database without worrying about
whether it'll change the allocator hash.

It also allows check_integrity() to avoid making an unnecessary commit
if the database is clean, which is nice with quick-repair. It means a
successful check_integrity() won't invalidate the allocator state table,
so you can still do a quick-repair afterwards