Ensure quick-repair commit on shutdown
Traditionally, redb has stored its allocator state in the region headers
and region tracker page, which are only valid after a clean shutdown.
Now that we also have the allocator state table (which works even after
a crash), it would be nice to use that exclusively.

This is the first step in that transition: making sure that after a clean
shutdown, the allocator state is saved to both places. Then, once enough
time has passed, we can remove all support for reading and writing the
old allocator state.

Even after the transition, any redb version will be able to open any
database. If the database has been opened at least once during the
transition period, it'll open cleanly; otherwise it'll require repair,
but it'll always work.