Close database file when Database is dropped
Previously the final flush was delayed until all open ReadTransactions
had been dropped, now it is done immediately.

Also, all operations will now return DatabaseClosed after the Database object
is dropped. This reduces the likelihood of data races on platforms that
do not support file locks. For example, the user may no longer open a
read transactions, close and re-open the Database, and then concurrently
write and read with the old transaction that is no longer associated
with the Database