Remove Durability::Eventual
The WriteTransaction::commit() API is
problematic when used with barrier sync, since there is no way to report
an error that might occur while writing out the data.

Applications that need this feature can implement it with a custom
StorageBackend. It was only implemented on macOS and called fcntl() with
F_BARRIERFSYNC. On all other platforms Durability::Eventual was
equivalent to Durability::Immediate