Fix a correctness issue with range queries
If a RangeBound where the start value was larger than the end was
passed to range(), drain_filter(), or drain() a non-empty range
could be returned. This would occur if the start and end values were
stored on separate pages (i.e. a sufficient condition is that 4KiB of
key-value pairs were stored in between them).

In the case of drain_filter() and drain() values would also have
been removed from the table, if the predicate indicated so