feat: support `wasm32-unknown-unknown` by disabling certain constructors
redb doesn't really care about the storage backend; it just needs
something that implements the `StorageBackend` trait. That was the
point of #707.

However, it doesn't actually build for `wasm32-unknown-unknown`,
because there are a bunch of convenience initializers which implicitly
instantiate a `FileBackend`, and of course that doesn't work on that
target.

This commit simply excludes those initializers where the target os
is unknown. As long as it is provided an appropriate `StorageBackend`,
redb will just do its thing; it doesn't need to care in that case
about how precisely the storage backend is implemented.