fix: use wasm-viable condvar implementation when targeting `wasm32-unknown-*`
The stdlib `Condvar` type panics at runtime if `.wait` is called in
`wasm32-unknown-unknown`. This commit replaces the only instance thereof with
an implementation which works on wasm, when compiling for that target
family and os.

Note that `#[cfg(target = "wasm32-unknown-unknown")]` is not legal syntax,
even though it would more concisely express our intent in this case.